fix: localStorage error

+++ pages/_app.js
@@ -3,16 +3,16 @@ import '../styles/globals.css'
 import Header from '../components/Header'
 import Footer from '../components/Footer'
 import jwtDecode from 'jwt-decode'
-import { useRouter } from 'next/router'
 
-let token = jwtDecode(localStorage.getItem('token'))
-//Check for expired token
-const currentTime = Date.now() / 1000;
-if (token.exp < currentTime) {
-  localStorage.clear()
-  router.push('/login')
+if (typeof window !== "undefined") {
+  let token = jwtDecode(localStorage.getItem('token'))
+  //Check for expired token
+  const currentTime = Date.now() / 1000;
+  if (token.exp < currentTime) {
+    localStorage.clear()
+    window.location.href = '/login'
+  }
 }
-
 function MyApp({ Component, pageProps }) {
 
   return (
